profile completed

+++ domain/src/main/java/com/paya/domain/models/remote/LinearChartRemoteModel.kt
@@ -3,7 +3,7 @@ package com.paya.domain.models.remote
 import com.google.gson.annotations.SerializedName
 
 data class LinearChartRemoteModel (
-	val data: List<Int>,
+	val data: List<Long>,
 	@SerializedName("start_date")
 	val startDate: String,
 	@SerializedName("end_date")
