+++ domain/src/main/java/com/paya/domain/models/repo/LinearChartRepoModel.kt
@@ -1,7 +1,7 @@
 package com.paya.domain.models.repo
 
 data class LinearChartRepoModel (
-	val data: List<Int>,
+	val data: List<Long>,
 	val startDate: String,
 	val endDate: String
 )
